+++ core/dbt/compilation.py
@@ -9,7 +9,7 @@ from dbt import flags
 from dbt.adapters.factory import get_adapter
 from dbt.clients import jinja
 from dbt.clients.system import make_directory
-from dbt.context.providers import generate_runtime_model
+from dbt.context.providers import generate_runtime_model_context
 from dbt.contracts.graph.manifest import Manifest, UniqueID
 from dbt.contracts.graph.compiled import (
     COMPILED_TYPES,
@@ -178,7 +178,7 @@ class Compiler:
         extra_context: Dict[str, Any],
     ) -> Dict[str, Any]:
 
-        context = generate_runtime_model(
+        context = generate_runtime_model_context(
             node, self.config, manifest
         )
         context.update(extra_context)