+++ core/dbt/config/runtime.py
@@ -253,7 +253,7 @@ class RuntimeConfig(Project, Profile, AdapterRequiredConfig):
     ) -> PathSet:
         for key, value in config.items():
             if isinstance(value, dict) and not key.startswith('+'):
-                self._get_v2_config_paths(value, path + (key,), paths)
+                self._get_config_paths(value, path + (key,), paths)
             else:
                 paths.add(path)
         return frozenset(paths)