+++ core/dbt/context/providers.py
@@ -11,7 +11,7 @@ from dbt.adapters.factory import (
     get_adapter, get_adapter_package_names, get_adapter_type_names
 )
 from dbt.clients import agate_helper
-from dbt.clients.jinja import get_rendered, MacroGenerator, MacroStack
+from dbt.clients.jinja import get_rendered, MacroGenerator, MacroStack, undefined_error
 from dbt.config import RuntimeConfig, Project
 from .base import contextmember, contextproperty, Var
 from .configured import FQNLookup
@@ -49,7 +49,7 @@ from dbt.exceptions import (
     wrapped_exports,
 )
 from dbt.config import IsFQNResource
-from dbt.logger import GLOBAL_LOGGER as logger  # noqa
+from dbt.logger import GLOBAL_LOGGER as logger, SECRET_ENV_PREFIX  # noqa
 from dbt.node_types import NodeType
 
 from dbt.utils import (
@@ -636,6 +636,7 @@ T = TypeVar('T')
 
 # Base context collection, used for parsing configs.
 class ProviderContext(ManifestContext):
+    # subclasses are MacroContext, ModelContext, TestContext
     def __init__(
         self,
         model,
@@ -1161,6 +1162,30 @@ class ProviderContext(ManifestContext):
         )
         raise CompilationException(msg)
 
+    @contextmember
+    def env_var(self, var: str, default: Optional[str] = None) -> str:
+        """The env_var() function. Return the environment variable named 'var'.
+        If there is no such environment variable set, return the default.
+
+        If the default is None, raise an exception for an undefined variable.
+        """
+        return_value = None
+        if var in os.environ:
+            return_value = os.environ[var]
+        elif default is not None:
+            return_value = default
+
+        if return_value is not None:
+            # Save the env_var value in the manifest and the var name in the source_file
+            if not var.startswith(SECRET_ENV_PREFIX) and self.model:
+                self.manifest.env_vars[var] = return_value
+                source_file = self.manifest.files[self.model.file_id]
+                source_file.env_vars.append(var)
+            return return_value
+        else:
+            msg = f"Env var required but not provided: '{var}'"
+            undefined_error(msg)
+
 
 class MacroContext(ProviderContext):
     """Internally, macros can be executed like nodes, with some restrictions:
@@ -1262,7 +1287,7 @@ class ModelContext(ProviderContext):
 
 
 # This is called by '_context_for', used in 'render_with_context'
-def generate_parser_model(
+def generate_parser_model_context(
     model: ManifestNode,
     config: RuntimeConfig,
     manifest: Manifest,
@@ -1279,7 +1304,7 @@ def generate_parser_model(
     return ctx.to_dict()
 
 
-def generate_generate_component_name_macro(
+def generate_generate_name_macro_context(
     macro: ParsedMacro,
     config: RuntimeConfig,
     manifest: Manifest,
@@ -1290,7 +1315,7 @@ def generate_generate_component_name_macro(
     return ctx.to_dict()
 
 
-def generate_runtime_model(
+def generate_runtime_model_context(
     model: ManifestNode,
     config: RuntimeConfig,
     manifest: Manifest,
@@ -1301,7 +1326,7 @@ def generate_runtime_model(
     return ctx.to_dict()
 
 
-def generate_runtime_macro(
+def generate_runtime_macro_context(
     macro: ParsedMacro,
     config: RuntimeConfig,
     manifest: Manifest,