+++ test/integration/068_partial_parsing_tests/test-files/env_var_model.sql
@@ -1,0 +1,1 @@
+select '{{ env_var('ENV_VAR_TEST') }}' as vartest