+++ test/unit/test_contracts_graph_parsed.py
@@ -122,7 +122,7 @@ def base_parsed_model_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Model),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -182,6 +182,7 @@ def basic_parsed_model_object():
         config=NodeConfig(),
         meta={},
         checksum=FileHash.from_contents(''),
+        created_at=1.0,
     )
 
 
@@ -190,7 +191,7 @@ def minimal_parsed_model_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Model),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -211,7 +212,7 @@ def complex_parsed_model_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Model),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -404,7 +405,7 @@ def basic_parsed_seed_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Seed),
         'path': '/root/seeds/seed.csv',
         'original_file_path': 'seeds/seed.csv',
@@ -477,7 +478,7 @@ def minimal_parsed_seed_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Seed),
         'path': '/root/seeds/seed.csv',
         'original_file_path': 'seeds/seed.csv',
@@ -497,7 +498,7 @@ def complex_parsed_seed_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Seed),
         'path': '/root/seeds/seed.csv',
         'original_file_path': 'seeds/seed.csv',
@@ -718,6 +719,8 @@ def patched_model_object():
 def test_patch_parsed_model(basic_parsed_model_object, basic_parsed_model_patch_object, patched_model_object):
     pre_patch = basic_parsed_model_object
     pre_patch.patch(basic_parsed_model_patch_object)
+    pre_patch.created_at = 1.0
+    patched_model_object.created_at = 1.0
     assert patched_model_object == pre_patch
 
 
@@ -745,7 +748,7 @@ def base_parsed_hook_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Operation),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -815,7 +818,7 @@ def complex_parsed_hook_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Operation),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -933,7 +936,7 @@ def minimal_parsed_schema_test_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Test),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -958,7 +961,7 @@ def basic_parsed_schema_test_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Test),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -1030,7 +1033,7 @@ def complex_parsed_schema_test_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Test),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -1372,7 +1375,7 @@ def basic_timestamp_snapshot_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Snapshot),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -1504,7 +1507,7 @@ def basic_check_snapshot_dict():
     return {
         'name': 'foo',
         'root_path': '/root/',
-        'created_at': 1,
+        'created_at': 1.0,
         'resource_type': str(NodeType.Snapshot),
         'path': '/root/x/path.sql',
         'original_file_path': '/root/path.sql',
@@ -1716,7 +1719,7 @@ class TestParsedMacro(ContractTestCase):
             'name': 'foo',
             'path': '/root/path.sql',
             'original_file_path': '/root/path.sql',
-            'created_at': 1,
+            'created_at': 1.0,
             'package_name': 'test',
             'macro_sql': '{% macro foo() %}select 1 as id{% endmacro %}',
             'root_path': '/root/',
@@ -1807,7 +1810,7 @@ def minimum_parsed_source_definition_dict():
         'root_path': '/root',
         'path': '/root/models/sources.yml',
         'original_file_path': '/root/models/sources.yml',
-        'created_at': 1,
+        'created_at': 1.0,
         'database': 'some_db',
         'schema': 'some_schema',
         'fqn': ['test', 'source', 'my_source', 'my_source_table'],
@@ -1828,7 +1831,7 @@ def basic_parsed_source_definition_dict():
         'root_path': '/root',
         'path': '/root/models/sources.yml',
         'original_file_path': '/root/models/sources.yml',
-        'created_at': 1,
+        'created_at': 1.0,
         'database': 'some_db',
         'schema': 'some_schema',
         'fqn': ['test', 'source', 'my_source', 'my_source_table'],
@@ -1884,7 +1887,7 @@ def complex_parsed_source_definition_dict():
         'root_path': '/root',
         'path': '/root/models/sources.yml',
         'original_file_path': '/root/models/sources.yml',
-        'created_at': 1,
+        'created_at': 1.0,
         'database': 'some_db',
         'schema': 'some_schema',
         'fqn': ['test', 'source', 'my_source', 'my_source_table'],
@@ -2039,7 +2042,7 @@ def minimal_parsed_exposure_dict():
         'root_path': '/usr/src/app',
         'original_file_path': 'models/something.yml',
         'description': '',
-        'created_at': 1,
+        'created_at': 1.0,
     }
 
 
@@ -2067,7 +2070,7 @@ def basic_parsed_exposure_dict():
         'description': '',
         'meta': {},
         'tags': [],
-        'created_at': 1,
+        'created_at': 1.0,
     }
 
 
@@ -2094,7 +2097,7 @@ def complex_parsed_exposure_dict():
     return {
         'name': 'my_exposure',
         'type': 'analysis',
-        'created_at': 1,
+        'created_at': 1.0,
         'owner': {
             'email': 'test@example.com',
             'name': 'A Name',